Move client extension send logic into their own functions
